refactor: memberId -> userId로 수정
- BE와 통일하기 위해 userId로 수정
