+++ src/app/member/[userId]/MemberPage.tsx
@@ -5,11 +5,11 @@ import { ApiEmotion } from "@/types/Emotion";
 import { useParams } from "react-router-dom";
 
 export default function MemberPage() {
-  const { memberId } = useParams();
+  const { userId } = useParams();
 
   return (
     <div className="flex flex-col gap-4 items-center">
-      <ProfileSummary memberId={Number(memberId)} />
+      <ProfileSummary userId={Number(userId)} />
       <div className="pt-2 flex flex-col gap-2.5 px-2 w-full">
         {/* TODO: 게시글 조회 추가 */}
         <PostCard key={`post-1`} postId={1}>
